philosopher

+++ cureses/minitalk/clientsrc/client.c
@@ -6,7 +6,7 @@
 /*   By: minhjang <minhjang@student.42seoul.kr>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/07/27 20:57:33 by minhjang          #+#    #+#             */
-/*   Updated: 2022/08/04 09:01:10 by minhjang         ###   ########.fr       */
+/*   Updated: 2022/08/04 09:03:15 by minhjang         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -21,7 +21,7 @@ int	main(int argc, char **argv)
 	if (argc != 3)
 		return (error_msg("You must enter pid and string"));
 	pid = pid_check(argv[1]);
-	if (pid == -1)
+	if (pid == (unsigned int)-1)
 		return (error_msg("pid range error"));
 	signal(SIGUSR1, receive_check_handler);
 	signal(SIGUSR2, error_catcher);
